requirements.txt
diffusers
accelerate
bitsandbytes==0.35.0
ftfy
gradio
safetensors
torchvision
transformers
triton
xformers
